+++ scripts/actions/target-actions.js
@@ -27,58 +27,20 @@ export const TARGET_ACTIONS = [
   // TODO ... Look at how a skillMod -2 affects the activation, as the Bolt power should activate on a 4 result, ignoring the -2 imposed by Dodge.
   //      ... So on a 4 or 5, it will cost the full Power Points to activate, but still fail to hit the target.
   {
-    id: "TARGET-HAS-DODGE-BOLT",
-    name: "Dodge vs Bolt",
-    button_name: "has Dodge vs Bolt",
-    //skillMod: "0",  //TODO this was so the power in the card Rolls section, but "0" seems to cause an error, so leave out for now.
-    and_selector: [
-      {
-        selector_type: "target_has_edge",
-        selector_value: "BRSW.EdgeName.Dodge", // NOTE This works because the global_actions.js expects ... value.includes("BRSW.EdgeName.")
-      },
-      {
-        selector_type: "item_type",
-        selector_value: "power"
-      },
-      {
-        selector_type: "item_name",
-        selector_value: "Bolt"
-      }
-    ],
-    defaultChecked: "on",
-    section: "attack",
-    group: "BRSW.Target",
-  },
-  {
-    id: "TARGET-HAS-DODGE-WEAPONS",
-    name: "Dodge vs Ranged",
-    button_name: "has Dodge vs Ranged",
+    id: "TARGET-HAS-DODGE",
+    name: "Dodge",
+    button_name: "has Dodge",
     skillMod: "-2",
+    ignoresArcaneActivation: true,
     and_selector: [
       {
         selector_type: "target_has_edge",
         selector_value: "BRSW.EdgeName.Dodge",
       },
       {
-        selector_type: "item_type",
-        selector_value: "weapon"
-      },
-      {
-        or_selector: [
-          {
-            selector_type: "skill",
-            selector_value: "BRSW.SkillName.UnskilledAttempt"
-          },
-          {
-            selector_type: "skill",
-            selector_value: "BRSW.SkillName.Shooting"
-          },
-          {
-            selector_type: "skill",
-            selector_value: "BRSW.SkillName.Athletics"
-          },
-        ],
-      },
+        selector_type: "item_is_weapon_or_bolt",
+        selector_value: "true"
+      }
     ],
     defaultChecked: "on",
     section: "attack",